IOHIDMouseAcceleration is an int32 on Monterey

+++ src/System/KillMacMouseAcceleration.c
@@ -40,8 +40,8 @@ typedef char BogusTypedef;  // work around ISO C warning about empty translation
 #include <IOKit/hidsystem/IOHIDParameter.h>
 #include <IOKit/hidsystem/event_status_driver.h>
 
-static const int64_t	kNoAcceleration			= -0x10000;
-static int64_t			gAccelerationBackup		= 0;
+static const int32_t	kNoAcceleration			= -0x10000;
+static int32_t			gAccelerationBackup		= 0;
 static bool				gAccelerationTainted	= false;
 
 void KillMacMouseAcceleration(void)
@@ -79,12 +79,12 @@ void KillMacMouseAcceleration(void)
 		ret = IOHIDSetParameter(handle, CFSTR(kIOHIDMouseAccelerationType), &kNoAcceleration, sizeof(kNoAcceleration));
 		if (ret != KERN_SUCCESS)
 		{
-			printf("%s: IOHIDSetParameter failed! Error %d. (Current accel = %lld)\n", __func__, (int)ret, gAccelerationBackup);
+			printf("%s: IOHIDSetParameter failed! Error %d. (Current accel = %d)\n", __func__, (int)ret, gAccelerationBackup);
 		}
 		else
 		{
 			gAccelerationTainted = true;
-			printf("%s: success. Was %lld, now %lld.\n", __func__, gAccelerationBackup, kNoAcceleration);
+			printf("%s: success. Was %d, now %d.\n", __func__, gAccelerationBackup, kNoAcceleration);
 		}
 	}
 	
@@ -95,7 +95,7 @@ void RestoreMacMouseAcceleration(void)
 {
 	if (!gAccelerationTainted)
 	{
-		printf("%s: Acceleration value not tainted (%lld).\n", __func__, gAccelerationBackup);
+		printf("%s: Acceleration value not tainted (%d).\n", __func__, gAccelerationBackup);
 		return;
 	}
 	
@@ -117,7 +117,7 @@ void RestoreMacMouseAcceleration(void)
 	else
 	{
 		gAccelerationTainted = false;
-		printf("%s: success. Restored %lld.\n", __func__, gAccelerationBackup);
+		printf("%s: success. Restored %d.\n", __func__, gAccelerationBackup);
 	}
 	
 	NXCloseEventStatus(handle);
